requirements.txt
streamlit==1.31.0
anthropic==0.9.0
openai==1.6.0
python-dotenv==1.0.0
httpx==0.26.0
asyncio==3.4.3
aiofiles==23.2.1
tenacity==8.2.3
pillow==10.1.0
tqdm==4.66.1
alive-progress==3.1.5
